# Conflicts:
#	core/src/components/UnifiedSearch/UnifiedSearchModal.vue

+++ apps/settings/src/components/AuthToken.vue
@@ -4,21 +4,23 @@
 -->
 
 <template>
-	<tr :class="['auth-token', { 'auth-token--wiping': wiping }]" :data-id="token.id">
+	<tr class="auth-token" :class="[{ 'auth-token--wiping': wiping }]" :data-id="token.id">
 		<td class="auth-token__name">
 			<NcIconSvgWrapper :path="tokenIcon" />
 			<div class="auth-token__name-wrapper">
-				<form v-if="token.canRename && renaming"
+				<form
+					v-if="token.canRename && renaming"
 					class="auth-token__name-form"
 					@submit.prevent.stop="rename">
-					<NcTextField ref="input"
-						:value.sync="newName"
+					<NcTextField
+						ref="input"
+						v-model="newName"
 						:label="t('settings', 'Device name')"
 						:show-trailing-button="true"
 						:trailing-button-label="t('settings', 'Cancel renaming')"
 						@trailing-button-click="cancelRename"
 						@keyup.esc="cancelRename" />
-					<NcButton :aria-label="t('settings', 'Save new name')" type="tertiary" native-type="submit">
+					<NcButton :aria-label="t('settings', 'Save new name')" variant="tertiary" type="submit">
 						<template #icon>
 							<NcIconSvgWrapper :path="mdiCheck" />
 						</template>
@@ -29,22 +31,26 @@
 			</div>
 		</td>
 		<td>
-			<NcDateTime class="auth-token__last-activity"
+			<NcDateTime
+				class="auth-token__last-activity"
 				:ignore-seconds="true"
 				:timestamp="tokenLastActivity" />
 		</td>
 		<td class="auth-token__actions">
-			<NcActions v-if="!token.current"
+			<NcActions
+				v-if="!token.current"
 				:title="t('settings', 'Device settings')"
 				:aria-label="t('settings', 'Device settings')"
 				:open.sync="actionOpen">
-				<NcActionCheckbox v-if="canChangeScope"
-					:checked="token.scope.filesystem"
-					@update:checked="updateFileSystemScope">
+				<NcActionCheckbox
+					v-if="canChangeScope"
+					:model-value="token.scope.filesystem"
+					@update:modelValue="updateFileSystemScope">
 					<!-- TODO: add text/longtext with some description -->
 					{{ t('settings', 'Allow filesystem access') }}
 				</NcActionCheckbox>
-				<NcActionButton v-if="token.canRename"
+				<NcActionButton
+					v-if="token.canRename"
 					icon="icon-rename"
 					@click.stop.prevent="startRename">
 					<!-- TODO: add text/longtext with some description -->
@@ -53,18 +59,21 @@
 
 				<!-- revoke & wipe -->
 				<template v-if="token.canDelete">
-					<template v-if="token.type !== 2">
-						<NcActionButton icon="icon-delete"
+					<template v-if="token.type !== TokenType.WIPING_TOKEN">
+						<NcActionButton
+							icon="icon-delete"
 							@click.stop.prevent="revoke">
 							<!-- TODO: add text/longtext with some description -->
 							{{ t('settings', 'Revoke') }}
 						</NcActionButton>
-						<NcActionButton icon="icon-delete"
+						<NcActionButton
+							icon="icon-delete"
 							@click.stop.prevent="wipe">
 							{{ t('settings', 'Wipe device') }}
 						</NcActionButton>
 					</template>
-					<NcActionButton v-else-if="token.type === 2"
+					<NcActionButton
+						v-else
 						icon="icon-delete"
 						:name="t('settings', 'Revoke')"
 						@click.stop.prevent="revoke">
@@ -73,54 +82,33 @@
 				</template>
 			</NcActions>
 		</td>
+		<AuthTokenDeleteDialog
+			v-if="deleteDialogOpen"
+			:token="token"
+			:open.sync="deleteDialogOpen"
+			@confirm="confirmDelete" />
 	</tr>
 </template>
 
 <script lang="ts">
 import type { PropType } from 'vue'
-import type { IToken } from '../store/authtoken'
+import type { IToken } from '../store/authtoken.ts'
 
-import { mdiCheck, mdiCellphone, mdiTablet, mdiMonitor, mdiWeb, mdiKeyOutline, mdiMicrosoftEdge, mdiFirefox, mdiGoogleChrome, mdiAppleSafari, mdiAndroid, mdiAppleIos } from '@mdi/js'
+import { mdiAndroid, mdiAppleIos, mdiAppleSafari, mdiCellphone, mdiCheck, mdiFirefox, mdiGoogleChrome, mdiKeyOutline, mdiMicrosoftEdge, mdiMonitor, mdiTablet, mdiWeb } from '@mdi/js'
+import { showConfirmation } from '@nextcloud/dialogs'
 import { translate as t } from '@nextcloud/l10n'
 import { defineComponent } from 'vue'
-import { TokenType, useAuthTokenStore } from '../store/authtoken.ts'
-
-import NcActions from '@nextcloud/vue/components/NcActions'
 import NcActionButton from '@nextcloud/vue/components/NcActionButton'
 import NcActionCheckbox from '@nextcloud/vue/components/NcActionCheckbox'
+import NcActions from '@nextcloud/vue/components/NcActions'
 import NcButton from '@nextcloud/vue/components/NcButton'
 import NcDateTime from '@nextcloud/vue/components/NcDateTime'
 import NcIconSvgWrapper from '@nextcloud/vue/components/NcIconSvgWrapper'
 import NcTextField from '@nextcloud/vue/components/NcTextField'
+import AuthTokenDeleteDialog from './AuthTokenDeleteDialog.vue'
+import { TokenType, useAuthTokenStore } from '../store/authtoken.ts'
+import { detect } from '../utils/userAgentDetect.ts'
 
-// When using capture groups the following parts are extracted the first is used as the version number, the second as the OS
-const userAgentMap = {
-	ie: /(?:MSIE|Trident|Trident\/7.0; rv)[ :](\d+)/,
-	// Microsoft Edge User Agent from https://msdn.microsoft.com/en-us/library/hh869301(v=vs.85).aspx
-	edge: /^Mozilla\/5\.0 \([^)]+\) AppleWebKit\/[0-9.]+ \(KHTML, like Gecko\) Chrome\/[0-9.]+ (?:Mobile Safari|Safari)\/[0-9.]+ Edge\/[0-9.]+$/,
-	// Firefox User Agent from https://developer.mozilla.org/en-US/docs/Web/HTTP/Gecko_user_agent_string_reference
-	firefox: /^Mozilla\/5\.0 \([^)]*(Windows|OS X|Linux)[^)]+\) Gecko\/[0-9.]+ Firefox\/(\d+)(?:\.\d)?$/,
-	// Chrome User Agent from https://developer.chrome.com/multidevice/user-agent
-	chrome: /^Mozilla\/5\.0 \([^)]*(Windows|OS X|Linux)[^)]+\) AppleWebKit\/[0-9.]+ \(KHTML, like Gecko\) Chrome\/(\d+)[0-9.]+ (?:Mobile Safari|Safari)\/[0-9.]+$/,
-	// Safari User Agent from http://www.useragentstring.com/pages/Safari/
-	safari: /^Mozilla\/5\.0 \([^)]*(Windows|OS X)[^)]+\) AppleWebKit\/[0-9.]+ \(KHTML, like Gecko\)(?: Version\/([0-9]+)[0-9.]+)? Safari\/[0-9.A-Z]+$/,
-	// Android Chrome user agent: https://developers.google.com/chrome/mobile/docs/user-agent
-	androidChrome: /Android.*(?:; (.*) Build\/).*Chrome\/(\d+)[0-9.]+/,
-	iphone: / *CPU +iPhone +OS +([0-9]+)_(?:[0-9_])+ +like +Mac +OS +X */,
-	ipad: /\(iPad; *CPU +OS +([0-9]+)_(?:[0-9_])+ +like +Mac +OS +X */,
-	iosClient: /^Mozilla\/5\.0 \(iOS\) (?:ownCloud|Nextcloud)-iOS.*$/,
-	androidClient: /^Mozilla\/5\.0 \(Android\) (?:ownCloud|Nextcloud)-android.*$/,
-	iosTalkClient: /^Mozilla\/5\.0 \(iOS\) Nextcloud-Talk.*$/,
-	androidTalkClient: /^Mozilla\/5\.0 \(Android\) Nextcloud-Talk.*$/,
-	// DAVx5/3.3.8-beta2-gplay (2021/01/02; dav4jvm; okhttp/4.9.0) Android/10
-	davx5: /DAV(?:droid|x5)\/([^ ]+)/,
-	// Mozilla/5.0 (U; Linux; Maemo; Jolla; Sailfish; like Android 4.3) AppleWebKit/538.1 (KHTML, like Gecko) WebPirate/2.0 like Mobile Safari/538.1 (compatible)
-	webPirate: /(Sailfish).*WebPirate\/(\d+)/,
-	// Mozilla/5.0 (Maemo; Linux; U; Jolla; Sailfish; Mobile; rv:31.0) Gecko/31.0 Firefox/31.0 SailfishBrowser/1.0
-	sailfishBrowser: /(Sailfish).*SailfishBrowser\/(\d+)/,
-	// Neon 1.0.0+1
-	neon: /Neon \d+\.\d+\.\d+\+\d+/,
-}
 const nameMap = {
 	edge: 'Microsoft Edge',
 	firefox: 'Firefox',
@@ -143,6 +131,7 @@ const nameMap = {
 export default defineComponent({
 	name: 'AuthToken',
 	components: {
+		AuthTokenDeleteDialog,
 		NcActions,
 		NcActionButton,
 		NcActionCheckbox,
@@ -151,29 +140,36 @@ export default defineComponent({
 		NcIconSvgWrapper,
 		NcTextField,
 	},
+
 	props: {
 		token: {
 			type: Object as PropType<IToken>,
 			required: true,
 		},
 	},
+
 	setup() {
 		const authTokenStore = useAuthTokenStore()
 		return { authTokenStore }
 	},
+
 	data() {
 		return {
 			actionOpen: false,
 			renaming: false,
 			newName: '',
 			oldName: '',
+			deleteDialogOpen: false,
 			mdiCheck,
+			TokenType,
 		}
 	},
+
 	computed: {
 		canChangeScope() {
 			return this.token.type === TokenType.PERMANENT_TOKEN
 		},
+
 		/**
 		 * Object ob the current user agent used by the token
 		 * This either returns an object containing user agent information or `null` if unknown
@@ -190,25 +186,16 @@ export default defineComponent({
 				}
 			}
 
-			for (const client in userAgentMap) {
-				const matches = this.token.name.match(userAgentMap[client])
-				if (matches) {
-					return {
-						id: client,
-						os: matches[2] && matches[1],
-						version: matches[2] ?? matches[1],
-					}
-				}
-			}
-
-			return null
+			return detect(this.token.name)
 		},
+
 		/**
 		 * Last activity of the token as ECMA timestamp (in ms)
 		 */
 		tokenLastActivity() {
 			return this.token.lastActivity * 1000
 		},
+
 		/**
 		 * Icon to use for the current token
 		 */
@@ -219,34 +206,35 @@ export default defineComponent({
 			}
 
 			switch (this.client?.id) {
-			case 'edge':
-				return mdiMicrosoftEdge
-			case 'firefox':
-				return mdiFirefox
-			case 'chrome':
-				return mdiGoogleChrome
-			case 'safari':
-				return mdiAppleSafari
-			case 'androidChrome':
-			case 'androidClient':
-			case 'androidTalkClient':
-				return mdiAndroid
-			case 'iphone':
-			case 'iosClient':
-			case 'iosTalkClient':
-				return mdiAppleIos
-			case 'ipad':
-				return mdiTablet
-			case 'davx5':
-				return mdiCellphone
-			case 'syncClient':
-				return mdiMonitor
-			case 'webPirate':
-			case 'sailfishBrowser':
-			default:
-				return mdiWeb
+				case 'edge':
+					return mdiMicrosoftEdge
+				case 'firefox':
+					return mdiFirefox
+				case 'chrome':
+					return mdiGoogleChrome
+				case 'safari':
+					return mdiAppleSafari
+				case 'androidChrome':
+				case 'androidClient':
+				case 'androidTalkClient':
+					return mdiAndroid
+				case 'iphone':
+				case 'iosClient':
+				case 'iosTalkClient':
+					return mdiAppleIos
+				case 'ipad':
+					return mdiTablet
+				case 'davx5':
+					return mdiCellphone
+				case 'syncClient':
+					return mdiMonitor
+				case 'webPirate':
+				case 'sailfishBrowser':
+				default:
+					return mdiWeb
 			}
 		},
+
 		/**
 		 * Label to be shown for current token
 		 */
@@ -266,6 +254,7 @@ export default defineComponent({
 			}
 			return name
 		},
+
 		/**
 		 * If the current token is considered for remote wiping
 		 */
@@ -273,11 +262,13 @@ export default defineComponent({
 			return this.token.type === TokenType.WIPING_TOKEN
 		},
 	},
+
 	methods: {
 		t,
 		updateFileSystemScope(state: boolean) {
 			this.authTokenStore.setTokenScope(this.token, 'filesystem', state)
 		},
+
 		startRename() {
 			// Close action (popover menu)
 			this.actionOpen = false
@@ -289,20 +280,37 @@ export default defineComponent({
 				this.$refs.input!.select()
 			})
 		},
+
 		cancelRename() {
 			this.renaming = false
 		},
+
 		revoke() {
 			this.actionOpen = false
+			this.deleteDialogOpen = true
+		},
+
+		confirmDelete() {
 			this.authTokenStore.deleteToken(this.token)
 		},
+
 		rename() {
 			this.renaming = false
 			this.authTokenStore.renameToken(this.token, this.newName)
 		},
-		wipe() {
+
+		async wipe() {
 			this.actionOpen = false
-			this.authTokenStore.wipeToken(this.token)
+			const confirmed = await showConfirmation({
+				name: t('settings', 'Confirm wipe'),
+				text: t('settings', 'Do you really want to wipe your data from this device?'),
+				labelConfirm: t('settings', 'Wipe device'),
+				labelReject: t('settings', 'Cancel'),
+				severity: 'warning',
+			})
+			if (confirmed) {
+				this.authTokenStore.wipeToken(this.token)
+			}
 		},
 	},
 })
