+++ apps/settings/src/store/authtoken.ts
@@ -1,3 +1,4 @@
+import axios from '@nextcloud/axios'
 /**
  * SPDX-FileCopyrightText: 2023 Nextcloud GmbH and Nextcloud contributors
  * SPDX-License-Identifier: AGPL-3.0-or-later
@@ -5,32 +6,19 @@
 import { showError } from '@nextcloud/dialogs'
 import { loadState } from '@nextcloud/initial-state'
 import { translate as t } from '@nextcloud/l10n'
-import { confirmPassword } from '@nextcloud/password-confirmation'
+import { addPasswordConfirmationInterceptors, confirmPassword, PwdConfirmationMode } from '@nextcloud/password-confirmation'
 import { generateUrl } from '@nextcloud/router'
 import { defineStore } from 'pinia'
-
-import axios from '@nextcloud/axios'
-import logger from '../logger'
-
-import '@nextcloud/password-confirmation/dist/style.css'
+import logger from '../logger.ts'
 
 const BASE_URL = generateUrl('/settings/personal/authtokens')
-
-const confirm = () => {
-	return new Promise(resolve => {
-		window.OC.dialogs.confirm(
-			t('settings', 'Do you really want to wipe your data from this device?'),
-			t('settings', 'Confirm wipe'),
-			resolve,
-			true,
-		)
-	})
-}
+addPasswordConfirmationInterceptors(axios)
 
 export enum TokenType {
 	TEMPORARY_TOKEN = 0,
 	PERMANENT_TOKEN = 1,
 	WIPING_TOKEN = 2,
+	ONETIME_TOKEN = 3,
 }
 
 export interface IToken {
@@ -71,34 +59,36 @@ export const useAuthTokenStore = defineStore('auth-token', {
 	actions: {
 		/**
 		 * Update a token on server
+		 *
 		 * @param token Token to update
 		 */
 		async updateToken(token: IToken) {
-			const { data } = await axios.put(`${BASE_URL}/${token.id}`, token)
+			const { data } = await axios.put(`${BASE_URL}/${token.id}`, token, { confirmPassword: PwdConfirmationMode.Strict })
 			return data
 		},
 
 		/**
 		 * Add a new token
+		 *
 		 * @param name The token name
 		 */
 		async addToken(name: string) {
 			logger.debug('Creating a new app token')
 
 			try {
-				await confirmPassword()
+				const { data } = await axios.post<ITokenResponse>(BASE_URL, { name, oneTime: true }, { confirmPassword: PwdConfirmationMode.Strict })
 
-				const { data } = await axios.post<ITokenResponse>(BASE_URL, { name })
 				this.tokens.push(data.deviceToken)
 				logger.debug('App token created')
 				return data
-			} catch (error) {
+			} catch {
 				return null
 			}
 		},
 
 		/**
 		 * Delete a given app token
+		 *
 		 * @param token Token to delete
 		 */
 		async deleteToken(token: IToken) {
@@ -107,7 +97,7 @@ export const useAuthTokenStore = defineStore('auth-token', {
 			this.tokens = this.tokens.filter(({ id }) => id !== token.id)
 
 			try {
-				await axios.delete(`${BASE_URL}/${token.id}`)
+				await axios.delete(`${BASE_URL}/${token.id}`, { confirmPassword: PwdConfirmationMode.Strict })
 				logger.debug('App token deleted')
 				return true
 			} catch (error) {
@@ -121,6 +111,7 @@ export const useAuthTokenStore = defineStore('auth-token', {
 
 		/**
 		 * Wipe a token and the connected device
+		 *
 		 * @param token Token to wipe
 		 */
 		async wipeToken(token: IToken) {
@@ -128,11 +119,6 @@ export const useAuthTokenStore = defineStore('auth-token', {
 
 			try {
 				await confirmPassword()
-
-				if (!(await confirm())) {
-					logger.debug('Wipe aborted by user')
-					return
-				}
 
 				await axios.post(`${BASE_URL}/wipe/${token.id}`)
 				logger.debug('App token marked for wipe', { token })
@@ -149,6 +135,7 @@ export const useAuthTokenStore = defineStore('auth-token', {
 
 		/**
 		 * Rename an existing token
+		 *
 		 * @param token The token to rename
 		 * @param newName The new name to set
 		 */
@@ -173,6 +160,7 @@ export const useAuthTokenStore = defineStore('auth-token', {
 
 		/**
 		 * Set scope of the token
+		 *
 		 * @param token Token to set scope
 		 * @param scope scope to set
 		 * @param value value to set
